Split up commands into packages
Changed remind to schedule.
Moved help command to util package.
Moved ttmtg command to players package.
Moved schedule command to schedule package.
Implemented the usage of the packages in the main cmds.js file.
