+++ config.json
@@ -1,3 +1,3 @@
 {
-    "prefix" : "+"
+    "prefix" : "!"
 }
