+++ src/cmds.js
@@ -3,127 +3,23 @@
 const config = require("../config.json");
 const logger = require("./logging.js").logger;
 
+const cmds_util = require('./util/cmds.js');
+const cmds_schedule = require('./schedule/cmds.js');
+const cmds_players = require('./players/cmds.js');
 
 /** the players bot commands */
 const commands = {
 
-    help: {
-        name: 'help',
-        usage: config.prefix + 'help [command]',
-        help: 'displays the command list with command summaries',
+    // util
+    help: cmds_util.help,
 
-        // args should be string[], message should be the discord Message as given in the message event
-        func: function (args, message) {
+    // players
+    ttmtg: cmds_players.ttmtg,
 
+    // schedule
+    schedule: cmds_schedule.schedule
 
-            logCommand(this.name, args); // logging
-            
-
-            // help only takes at most one argument
-            if (args.length > 1) {
-                sendUsage(this.name, message);
-                return;
-            }
-
-
-            // help with one argument
-            if (args.length === 1) {
-
-                // argument must be a valid command
-                if (commands[args[0]] === undefined) {
-
-                    switch (args[0]) {
-
-                        case "me": // a nice little response?
-                            message.channel.send(`Anything you need, I'm always here ${message.member.displayName}.`);
-                            break;
-
-                        default: // people need to enter a command that we know
-
-                            unknownCommand(`help ${args[0]}`, [], message);
-                            break;
-
-                    }
-
-                    // display the command usage
-                } else sendUsage(args[0], message);
-                
-
-                // help with no argument
-            } else {
-
-                let msg = "Commands:"
-
-                // build help message
-                for (let key in commands)
-                    msg += "\n" + commands[key].name + " - " + commands[key].help;
-                
-                message.channel.send(msg);
-            }
-        }
-    },
-
-    ttmtg: {
-        name: "ttmtg",
-        usage: config.prefix + 'ttmtg',
-        help: 'displays the time until the next RIT Players meeting',
-        func: function (args, message) {
-
-            logCommand(this.name, args); // logging
-
-            let d = new Date(); // 'now'
-            let m = new Date(); // to be set to next meeting time
-
-            m.setDate(m.getDate() + ((7 - m.getDay()) % 7 + 2) % 7); // get date of next Tuesday (m will stay a Tuesday if it is already one)
-            m.setHours(20, 0, 0, 0); // meetings are at 8p
-
-            // if the time is after 8p on Tuesday, we need to move to the actual next Tuesday
-            if (d > m) m.setDate(m.getDate() + (7 - m.getDay()) % 7 + 2);
-
-            // get the difference in the dates
-            let ms = m - d;
-
-            // convert ms to days, hours, minutes, and seconds
-            let s = parseInt(Math.floor(ms / 1000)); // seconds
-            let min = parseInt(Math.floor(s / 60)); // minutes
-            let h = parseInt(Math.floor(min / 60)); // hours
-            let day = parseInt(Math.floor(h / 24)); // days
-            s %= 60;
-            min %= 60;
-            h %= 24;
-
-            message.channel.send(`${day} days, ${h} hours, ${min} minutes, and ${s} seconds until the next Players meeting.`);
-        }
-    },
-
-    reminders: {
-
-        name: "reminders",
-        usage: config.prefix + 'reminders',
-        help: 'lists the current reminders for this server',
-        func: function (args, message) {
-
-            logCommand(this.name, args); // logging
-
-            // todo
-
-        }
-
-    },
-
-    remind: {
-
-        name: "remind",
-        usage: config.prefix + 'remind ()', // todo
-        help: 'add a reminder to this server',
-        func: function (args, message) {
-
-            logCommand(this.name, args); // logging
-
-            // todo
-
-        }
-    }
+    // MORE COMMANDS GET ADDED HERE
 
 };
 
@@ -133,10 +29,10 @@ const commands = {
  * @param {string} cmd      the command
  * @param {string[]} args   the command arguments
  */
-function logCommand(cmd, args) {
+exports.logCommand = (cmd, args) => {
     logger.info(`"${cmd}" command with args:`);
     logger.info(args);
-}
+};
 
 
 /**
@@ -145,7 +41,7 @@ function logCommand(cmd, args) {
  * @param {string[]} args   the command arguments
  * @param {Message} message the message that the command was sent in
  */
-function unknownCommand(cmd, args, message) {
+exports.unknownCommand = (cmd, args, message) => {
 
     // logging
     logger.info(`Unknown command "${cmd}" with args:`);
@@ -154,7 +50,7 @@ function unknownCommand(cmd, args, message) {
     // unknown command message (todo?)
     message.channel.send(`I don't know '${cmd}'. Try running '${config.prefix}help'`);
 
-}
+};
 
 
 /**
@@ -162,11 +58,23 @@ function unknownCommand(cmd, args, message) {
  * @param {string} cmd      the command
  * @param {Message} message the message to respond to
  */
-function sendUsage(cmd, message) {
-    
-    message.channel.send(`Usage of ${cmd} ((required) [optional]): ${commands[cmd].usage}`); // todo
+exports.sendUsage = (cmd, message) => {
 
-}
+    message.channel.send(`Usage of ${cmd} ((required) [optional]): ${config.prefix}${commands[cmd].usage}`); // todo
+
+};
+
+
+/**
+ * Get all the bot commands as an object.
+ * @return the object
+ */
+exports.getCommands = () => {
+
+    // so that commands implemented not here can see other commands info without importing them
+    return commands;
+
+};
 
 
 /**
@@ -175,11 +83,11 @@ function sendUsage(cmd, message) {
  * @param {string[]} args   the command arguments
  * @param {Message} message the message that the command was sent in
  */
-exports.run = function (cmd, args, message) {
+exports.run = (cmd, args, message) => {
 
     // run the command if it is known
     if (commands[cmd]) commands[cmd].func(args, message);
 
-    else unknownCommand(cmd, args, message);
+    else exports.unknownCommand(cmd, args, message);
 
 };
